rename 'tensorflow/neuralNetworks ' #1256
rename it as 'tensorflow_nnet'
