+++ egs/wsj/s5/steps/libs/tensorflow/tensorflow_nnet/nnet.py
